[CHORE] 유저 정보를 공용컴포넌트화
